Fix FFI boundary

+++ pinkytwirl-rs/src/engine.rs
@@ -227,7 +227,7 @@ impl PinkyTwirlEngine {
 
     pub fn macos_handle_key_event(
         &mut self,
-        key_code: u16,
+        key_code: i64,
         down: bool,
         shift: bool,
         ctrl: bool,
